Error specificity when checking for continuous.dat
Added an additional ValueError for the event that `_read_analog_signals` finds multiple .dat files.
Future edits code could optionally default to using the one that contained 'continuous' in the filename. 
Found a need for this in my own case where user stored another dat file in the same place.